+++ springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/api/dto/OrderDTO.java
@@ -1,0 +1,4 @@
+package com.hello.chain.project.biz.api.dto;
+
+public class OrderDTO {
+}
