+++ springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/api/service/OrderService.java
@@ -1,0 +1,4 @@
+package com.hello.chain.project.biz.api.service;
+
+public class OrderService {
+}
