--- springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/dao/user/UserMapper.java
@@ -1,4 +1,0 @@
-package com.hello.chain.project.biz.dao.user;
-
-public class UserMapper {
-}
