+++ springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/dependency/ai/AiRPCManager.java
@@ -1,0 +1,4 @@
+package com.hello.chain.project.biz.dependency.ai;
+
+public class AiRPCManager {
+}
