+++ springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/dependency/data/DataHTTPManager.java
@@ -1,0 +1,4 @@
+package com.hello.chain.project.biz.dependency.data;
+
+public class DataHTTPManager {
+}
