--- springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/domain/entity/order/convert/UserConvert.java
@@ -1,4 +1,0 @@
-package com.hello.chain.project.biz.domain.entity.order.convert;
-
-public class UserConvert {
-}
