--- springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/domain/model/user/UserDO.java
@@ -1,4 +1,0 @@
-package com.hello.chain.project.biz.domain.model.user;
-
-public class UserDO {
-}
