+++ springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/domain/order/bo/OrderBo.java
@@ -1,0 +1,4 @@
+package com.hello.chain.project.biz.domain.order.bo;
+
+public class OrderBo {
+}
