+++ springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/domain/order/query/UserQuery.java
@@ -1,0 +1,4 @@
+package com.hello.chain.project.biz.domain.order.query;
+
+public class UserQuery {
+}
