+++ springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/domain/order/service/BookService.java
@@ -1,0 +1,7 @@
+package com.hello.chain.project.biz.domain.order.service;
+
+/**
+ * 数据与操作合并服务
+ */
+public class BookService {
+}
