+++ springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/infrastructure/dao/book/BookMapper.java
@@ -1,0 +1,4 @@
+package com.hello.chain.project.biz.infrastructure.dao.book;
+
+public class BookMapper {
+}
