+++ springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/infrastructure/dao/user/UserMapper.java
@@ -1,0 +1,4 @@
+package com.hello.chain.project.biz.infrastructure.dao.user;
+
+public class UserMapper {
+}
