+++ springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/infrastructure/model/book/BookDO.java
@@ -1,0 +1,4 @@
+package com.hello.chain.project.biz.infrastructure.model.book;
+
+public class BookDO {
+}
