+++ springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/infrastructure/model/user/UserDO.java
@@ -1,0 +1,4 @@
+package com.hello.chain.project.biz.infrastructure.model.user;
+
+public class UserDO {
+}
