+++ springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/manager/OrderManager.java
@@ -1,0 +1,4 @@
+package com.hello.chain.project.biz.manager;
+
+public interface OrderManager {
+}
