+++ springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/manager/impl/OrderManagerImpl.java
@@ -1,0 +1,7 @@
+package com.hello.chain.project.biz.manager.impl;
+
+
+import com.hello.chain.project.biz.manager.OrderManager;
+
+public class OrderManagerImpl implements OrderManager {
+}
