--- springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/manager/open/ai/AiApiManager.java
@@ -1,4 +1,0 @@
-package com.hello.chain.project.biz.manager.open.ai;
-
-public class AiApiManager {
-}
