--- springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/manager/open/data/DataApiManager.java
@@ -1,4 +1,0 @@
-package com.hello.chain.project.biz.manager.open.data;
-
-public class DataApiManager {
-}
