--- springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/manager/order/impl/OrderManagerImpl.java
@@ -1,7 +1,0 @@
-package com.hello.chain.project.biz.manager.order.impl;
-
-
-import com.hello.chain.project.biz.manager.order.OrderManager;
-
-public class OrderManagerImpl implements OrderManager {
-}
