--- springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/share/dto/BookDTO.java
@@ -1,4 +1,0 @@
-package com.hello.chain.project.biz.share.dto;
-
-public class BookDTO {
-}
