--- springboot-project/springboot-biz/src/main/java/com/hello/chain/project/biz/share/service/BookService.java
@@ -1,4 +1,0 @@
-package com.hello.chain.project.biz.share.service;
-
-public class BookService {
-}
